+++ src/app.ts
@@ -4,9 +4,11 @@ import express, {
   urlencoded,
   Response as ExResponse,
   Request as ExRequest,
+  ErrorRequestHandler,
 } from "express";
 import { RegisterRoutes } from "../generated/routes";
 import swaggerUi from "swagger-ui-express";
+import { ValidateError } from "tsoa";
 
 export const app = express();
 
@@ -20,6 +22,7 @@ app.use(json());
 
 RegisterRoutes(app);
 
+// Adding Swagger UI as devtool
 app.use(
   "/docs",
   swaggerUi.serve,
@@ -27,3 +30,31 @@ app.use(
     res.send(swaggerUi.generateHTML(await import("../generated/swagger.json")));
   }
 );
+
+// Not Found 404 errors:
+app.use(function notFoundHandler(_req, res: ExResponse) {
+  res.status(404).send({
+    message: "Not Found",
+  });
+});
+
+// Catch-all error handling
+const errorHandler: ErrorRequestHandler = (err, req, res, next) => {
+  if (err instanceof ValidateError) {
+    console.warn(`Caught Validation Error for ${req.path}:`, err.fields);
+    res.status(422).json({
+      message: "Validation Failed",
+      details: err?.fields,
+    });
+    return;
+  }
+  if (err instanceof Error) {
+    res.status(500).json({
+      message: "Internal Server Error",
+    });
+    return;
+  }
+
+  next();
+};
+app.use(errorHandler);